requirements.txt
alembic==1.5.5
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2020.12.5
charset-normalizer==2.0.7
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
google-api-core==2.2.1
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.26.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
uritemplate==4.1.1
urllib3==1.26.7
Wand==0.6.5
Werkzeug==1.0.1
wincertstore==0.2
